+++ packages/vue-router/src/lazyRouteComponent.tsx
@@ -47,7 +47,10 @@ export function lazyRouteComponent<
       error = undefined
       loadPromise = importer()
         .then((res) => {
-          loadPromise = undefined
+          if (typeof document !== 'undefined') {
+            loadPromise = undefined
+            ;(lazyComp as any).preload = undefined
+          }
           comp = res[exportName ?? 'default']
           return comp
         })
@@ -66,7 +69,6 @@ export function lazyRouteComponent<
 
     return loadPromise
   }
-
   // Create a lazy component wrapper using defineComponent so it works in Vue SFC templates
   const lazyComp = Vue.defineComponent({
     name: 'LazyRouteComponent',