+++ packages/react-start/CHANGELOG.md
@@ -1,5 +1,54 @@
 # @tanstack/react-start
 
+## 1.168.46
+
+### Patch Changes
+
+- Updated dependencies [[`0fdf9ff`](https://github.com/TanStack/router/commit/0fdf9ff16fc532de9fc131d18df167ebd8038720)]:
+  - @tanstack/react-router@1.170.29
+  - @tanstack/react-start-client@1.168.27
+  - @tanstack/react-start-rsc@0.1.45
+  - @tanstack/react-start-server@1.167.34
+  - @tanstack/start-client-core@1.170.24
+  - @tanstack/start-plugin-core@1.171.36
+  - @tanstack/start-server-core@1.169.28
+
+## 1.168.45
+
+### Patch Changes
+
+- Updated dependencies [[`0c25a7b`](https://github.com/TanStack/router/commit/0c25a7b43b5331ee6ebf94d50e9793002783adab), [`31882c7`](https://github.com/TanStack/router/commit/31882c7fa87debef236228831655cb112c20ce90)]:
+  - @tanstack/react-router@1.170.28
+  - @tanstack/react-start-client@1.168.26
+  - @tanstack/react-start-rsc@0.1.44
+  - @tanstack/react-start-server@1.167.33
+  - @tanstack/start-client-core@1.170.23
+  - @tanstack/start-plugin-core@1.171.35
+  - @tanstack/start-server-core@1.169.27
+
+## 1.168.44
+
+### Patch Changes
+
+- Updated dependencies [[`b0817f5`](https://github.com/TanStack/router/commit/b0817f50cc639ae50b6c30cd11c5adbf7eed028c)]:
+  - @tanstack/react-router@1.170.27
+  - @tanstack/react-start-client@1.168.25
+  - @tanstack/react-start-rsc@0.1.43
+  - @tanstack/react-start-server@1.167.32
+
+## 1.168.43
+
+### Patch Changes
+
+- Updated dependencies [[`f500760`](https://github.com/TanStack/router/commit/f5007607d62e932b8df19de866cdc6bddcff8db3), [`1aafca9`](https://github.com/TanStack/router/commit/1aafca9b45d24f8f0dba78e716fc9582346c6ca1), [`aa10b65`](https://github.com/TanStack/router/commit/aa10b6589eeda215f5a60ea0af95a3101d8c414e)]:
+  - @tanstack/react-router@1.170.26
+  - @tanstack/react-start-client@1.168.24
+  - @tanstack/react-start-rsc@0.1.42
+  - @tanstack/react-start-server@1.167.31
+  - @tanstack/start-client-core@1.170.22
+  - @tanstack/start-plugin-core@1.171.34
+  - @tanstack/start-server-core@1.169.26
+
 ## 1.168.42
 
 ### Patch Changes