+++ packages/router-cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tanstack/router-cli
 
+## 1.167.30
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.30
+
+## 1.167.29
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.29
+
+## 1.167.28
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-generator@1.167.28
+
 ## 1.167.27
 
 ### Patch Changes