+++ packages/router-vite-plugin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @tanstack/router-vite-plugin
 
+## 1.167.32
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-plugin@1.168.32
+
+## 1.167.31
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-plugin@1.168.31
+
+## 1.167.30
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/router-plugin@1.168.30
+
 ## 1.167.29
 
 ### Patch Changes